feat: modify the identifier in notification and related code to `userId`, and unify `userId` as a long data type

+++ src/main/java/com/github/learndifferent/mtm/mapper/BookmarkMapper.java
@@ -191,7 +191,7 @@ public interface BookmarkMapper {
      * @param bookmarkId ID of the bookmark
      * @return user ID of the bookmark owner
      */
-    Integer getBookmarkOwnerUserId(int bookmarkId);
+    Long getBookmarkOwnerUserId(int bookmarkId);
 
     /**
      * Get visited bookmarks
